requirements.txt
Click
dlib
face-recognition
face-recognition-models
numpy
Pillow
pixelsort